requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.34.2
requests==2.31.0
PyMuPDF==1.23.8
python-docx==1.1.0
markdown==3.5.1
weasyprint==60.1
werkzeug==3.0.1